% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_icnarc_site_description.R
\name{add_icnarc_site_description}
\alias{add_icnarc_site_description}
\title{add_icnarc_site_description}
\usage{
add_icnarc_site_description(main_df, cmp_code_col, new_col_name = "")
}
\arguments{
\item{main_df}{Dataframe to update}

\item{cmp_code_col}{Name of column containing ICNARC CMP codes}

\item{new_col_name}{Name of new column}
}
\value{
The updated dataframe
}
\description{
Given a dataframe with a column of ICNARC CMP codes, add_icnarc_site_description()
will add a new column containing the site descriptions
}
