% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_icnarc_cmp_format.R
\name{check_icnarc_cmp_format}
\alias{check_icnarc_cmp_format}
\title{check_icnarc_cmp_format}
\usage{
check_icnarc_cmp_format(df, column_name, strict = FALSE)
}
\arguments{
\item{strict}{If TRUE, will stop, otherwise returns a boolean vector indicating correct formatting}
}
\value{
A boolean vector indicating correct formatting
}
\description{
Checks the correct formatting of column of ICNARC CMP codes in a dataframe.
}
