% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateify_filename.R
\name{dateify_filename}
\alias{dateify_filename}
\title{dateify_filename}
\usage{
dateify_filename(filename)
}
\arguments{
\item{filename}{}
}
\value{
String of decorated filename
}
\description{
dateify_filename adds underscore + date(as YYYYMMDD) before the extension of a filename
}
