% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateifyFilename.R
\name{dateifyFilename}
\alias{dateifyFilename}
\title{dateifyFilename}
\usage{
dateifyFilename(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
dateifyFilename adds underscore + date(as YYYYMMDD) before the extension of a filename
}
