% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egof_table.R
\name{egof_table}
\alias{egof_table}
\title{egof_table}
\usage{
egof_table(glm_fit, n_quantiles = 10)
}
\arguments{
\item{glm_fit}{Fiited GLM model}

\item{n_quantiles}{Number of quantiles to group by}
}
\value{
data frame of observed and expected outcome counts
}
\description{
A function to ape the "egof table" of STATA
Pass a fitted glm object and specify the quantiles to group predicted probabilities
by, and returns tabulated observed and expected outcome counts.
Halfway to a Hosmer-Lemeshow test
}
\details{
Requires dplyr
}
