% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survdiff_pvalue.R
\name{get_survdiff_pvalue}
\alias{get_survdiff_pvalue}
\title{get_survdiff_pvalue gets p values from survdiff objects}
\usage{
get_survdiff_pvalue(survdiffobj)
}
\arguments{
\item{survdiffobj}{Object returned from survival::survdiff()}
}
\value{
p
}
\description{
get_survdiff_pvalue gets p values from survival::survdiff() objects.
As per https://stat.ethz.ch/pipermail/r-help/2001-November/016742.html
}
\details{
If things are not readily available in R it is always good to pause and 
reflect if there might be a good reason.
--Brian D. Ripley (about how to get t statistics for arima())
R-help March 2013
}
