% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurvdiffPvalue.R
\name{getSurvdiffPvalue}
\alias{getSurvdiffPvalue}
\title{getSurvdiffPvalue gets p values from survdiff objects}
\usage{
getSurvdiffPvalue(survdiffobj)
}
\arguments{
\item{survdiffobj}{Object returned from survival::survdiff()}
}
\value{
p
}
\description{
getSurvdiffPvalue gets p values from survival::survdiff() objects.
As per https://stat.ethz.ch/pipermail/r-help/2001-November/016742.html
}
