% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_diagnostic_ggplots.R
\name{lm_diagnostic_ggplots}
\alias{lm_diagnostic_ggplots}
\title{lm_diagnostic_ggplots}
\usage{
lm_diagnostic_ggplots(model)
}
\arguments{
\item{model}{lm() object}
}
\value{
A list of 6 ggplot objects
}
\description{
Returns six ggplots depicting various diagnostics for the lm model passed.
}
