% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make2part_SQL_query_with_list.R
\name{make2part_SQL_query_with_list}
\alias{make2part_SQL_query_with_list}
\title{make2part_SQL_query_with_list a SQL query helper}
\usage{
make2part_SQL_query_with_list(sql_frags, ids, quotes = TRUE)
}
\arguments{
\item{sql_frags}{Vector of 2 SQL fragments}

\item{ids}{Vector of IDs to comma separate and concatenate}

\item{quotes}{Quote each id, default TRUE}
}
\value{
full_query
}
\description{
make2part_SQL_query_with_list builds a query from 2 SQL fragments and a list of ids.
The first part of an ICIP SQL query is concatenated with a comma
separated list of ids and terminated with the second SQL fragment.
}
\details{
NOTE: Would probably just do this with glue now
}
