% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make2partSQLQueryWithList.R
\name{make2partSQLQueryWithList}
\alias{make2partSQLQueryWithList}
\title{make2partSQLQueryWithList a SQL query helper}
\usage{
make2partSQLQueryWithList(sql_frags, ids, quotes = TRUE)
}
\arguments{
\item{sql_frags}{Vector of 2 SQL fragments}

\item{ids}{Vector of IDs to comma separate and concatenate}

\item{quotes}{Quote each id, default TRUE}
}
\value{
full_query
}
\description{
make2partSQLQueryWithList builds a query from 2 SQL fragments and a list of ids.
The first part of an ICIP SQL query is concatenated with a comma
separated list of ids and terminated with the second SQL fragment
}
