% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_iqr_string.R
\name{median_iqr_string}
\alias{median_iqr_string}
\title{median_iqr_string returns a formatted string detailing median and IQR}
\usage{
median_iqr_string(x, na.rm = T, probs = c(0.25, 0.75), digits = 2)
}
\arguments{
\item{x}{A vector of numeric values to be evaluated}

\item{na.rm}{logical; if true, any NA and NaN's are removed from x before the quantiles are computed.}

\item{probs}{numeric vector of probabilities with values in [0,1], default 0.25 and 0.75, only first two probs will be reported as range}

\item{digits}{numeric number of digits for rounding, default=2}
}
\value{
result_string
}
\description{
median_iqr_string returns a formatted string detailing median and IQR
}
