% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyprint_est_ci.R
\name{prettyprint_est_ci}
\alias{prettyprint_est_ci}
\title{prettyprint_est_ci}
\usage{
prettyprint_est_ci(est, ll, ul, digits = 2, ci_pc = "95")
}
\arguments{
\item{est}{Estimate}

\item{ll}{Confidence interval lower limit}

\item{ul}{Confidence interval upper limit}

\item{digits}{Digits to round to}

\item{ci_pc}{Confidence interval percentage (default: 95)}
}
\value{
A string
}
\description{
Returns a string formatting an estimate and confidence interval bounds
}
