% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqladder.R
\name{qqladder_df}
\alias{qqladder_df}
\title{qqladder_df}
\usage{
qqladder_df(data)
}
\arguments{
\item{data}{Vector of values for transformation}
}
\description{
Returns long dataframe of transformed data ordered by transformation and value
As per stata qladder function
}
\examples{
df <- qqladder_df(rnorm(100, 10, 1))
}
