% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_dataframe.r
\name{table_to_dataframe}
\alias{table_to_dataframe}
\title{table_to_dataframe Expand a table object of counts to a dataframe}
\usage{
table_to_dataframe(t)
}
\arguments{
\item{t}{Table object}
}
\value{
dataframe of groups and outcomes
}
\description{
Expands a table object of counts to a dataframe, ie an inverse table() function.
Maybe should have worked harder on the google fu and stack overflow beforehand...
}
