% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_elem_not_numeric.R
\name{which_elem_not_numeric}
\alias{which_elem_not_numeric}
\title{which_elem_not_numeric}
\usage{
which_elem_not_numeric(x)
}
\arguments{
\item{x}{A vector of non numeric values}
}
\value{
A vector of indices
}
\description{
which_elem_not_numeric returns the indices of a vector that cannot be cast as numeric types.
That is elements of a vector that would raise 'NAs introduced by coercion' when as.numeric() applied to vector.
Warnings are supressed.
}
\examples{

which_elem_not_numeric(c(1, 2, 3, NA_character_, 'fred'))
[1] 4 5
}
