% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichElemNotNumeric.R
\name{whichElemNotNumeric}
\alias{whichElemNotNumeric}
\title{whichElemNotNumeric}
\usage{
whichElemNotNumeric(x)
}
\arguments{
\item{x}{A vector of non numeric values}
}
\value{
A vector of indices
}
\description{
whichElemNotNumeric returns the indices of a vector that cannot be cast as numeric types.
That is elements of a vector that would raise 'NAs introduced by coercion' when as.numeric() applied to vector.
Warnings are supressed.
}
\examples{

whichElemNotNumeric(c(1, 2, 3, NA_character_, 'fred'))
[1] 4 5
}
