% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket_and_quote.R
\name{bracket_and_quote}
\alias{bracket_and_quote}
\title{bracket_and_quote}
\usage{
bracket_and_quote(string_vec)
}
\arguments{
\item{string_vec}{}
}
\value{

}
\description{
bracket_and_quote takes a string vector, wraps each string in single quotes,
collapses all the strings into a comma separated version wrapped in () brackets.
}
