% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyGasLabels.R
\name{classifyGasLabels}
\alias{classifyGasLabels}
\title{classifyGasLabels classifies raw blood gas sample type labels}
\usage{
classifyGasLabels(labels)
}
\arguments{
\item{labels}{}
}
\value{
label_classifications
}
\description{
classifyGasLabels classifies vectors of raw blood gas sample type labels.
A vector of the following classifications is returned.
}
\details{
* ARETERIAL
* VENOUS
* MIXED_VENOUS
* PRE_ECMO
* POST_ECMO
* UNKNOWN
}
