% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bakenzua_clean.R
\name{theme_bakenzua_clean}
\alias{theme_bakenzua_clean}
\title{theme_bakenzua_clean}
\usage{
theme_bakenzua_clean(...)
}
\value{

}
\description{
ggplot defaults
}
\examples{
ggplot2::theme_set(theme_bakenzua_clean())
}
